airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
boto3==1.42.68
pyspark==4.1.1
delta-spark==4.1.0
great-expectations==0.18.21
dbt-core==1.11.7
dbt-duckdb==1.10.1


